+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - Change resample method to use default Hanning window (was not using a window previously).
  - Define new Metrics class to improving handling of intensity metrics.
  - Define new MetricsXML class to isolate the conversion of metrics into an XML.
+ - Define new Rupture class which adds rupture model information to workspace file.
  - Refactor StationSummary into multiple separate classes:
    - WaveformMetric, WaveformMetricList, WaveformMetricCollection, WaveformMetricsXML
    - StationMetric, StationMetricCollection